+++ packages/gatsby/src/templates/article.js
@@ -23,23 +23,37 @@ function getGitPageUrl(postAbsolutePath) {
 }
 
 // eslint-disable-next-line arca/no-default-export
-export default function Template({data, pageContext: {category}}) {
-  const {allMarkdownRemark, markdownRemark} = data;
-  const {frontmatter, html, fileAbsolutePath} = markdownRemark;
+export default function Template({data}) {
+  const {allMdx, mdx} = data;
+  const {frontmatter, body, fileAbsolutePath} = mdx;
+
+  const orderedItems = [];
+  const items = [];
+
+  for (const {node} of allMdx.edges) {
+    const {path, title, order} = node.frontmatter;
+    const item = {
+      to: path,
+      name: title,
+    };
+
+    if (typeof order === `number`) {
+      orderedItems[order - 1] = item;
+    } else {
+      items.push(item);
+    }
+  }
 
   return <>
     <Global styles={GlobalStyleOverrides} />
-    <LayoutContentNav items={allMarkdownRemark.edges.map(({node}) => ({
-      to: node.frontmatter.path,
-      name: node.frontmatter.title,
-    }))}>
+    <LayoutContentNav items={orderedItems.concat(items)}>
       <SEO
         title={frontmatter.title}
         description={frontmatter.description}
         keywords={[`package manager`, `yarn`, `yarnpkg`, frontmatter.path.split(`/`).reverse()[0]]}
       />
       <PrerenderedMarkdown title={frontmatter.title} editUrl={getGitPageUrl(fileAbsolutePath)}>
-        {html}
+        {body}
       </PrerenderedMarkdown>
     </LayoutContentNav>
   </>;
@@ -47,7 +61,7 @@ export default function Template({data, pageContext: {category}}) {
 
 export const pageQuery = graphql`
   query($path: String!, $category: String) {
-    allMarkdownRemark(
+    allMdx(
       filter: {frontmatter: {category: {eq: $category}, hidden: {ne: true}}}
       sort: {order: ASC, fields: [frontmatter___title]}
     ) {
@@ -56,12 +70,14 @@ export const pageQuery = graphql`
           frontmatter {
             path
             title
+            order
           }
         }
       }
     }
-    markdownRemark(frontmatter: {category: {eq: $category}, path: {eq: $path}}) {
-      html
+
+    mdx(frontmatter: {category: {eq: $category}, path: {eq: $path}}) {
+      body
       fileAbsolutePath
       frontmatter {
         path