+++ packages/yarnpkg-core/rollup.config.js
@@ -2,6 +2,8 @@ import cjs                  from '@rollup/plugin-commonjs';
 import resolve              from '@rollup/plugin-node-resolve';
 import path                 from 'path';
 import esbuild              from 'rollup-plugin-esbuild';
+import {terser}             from 'rollup-plugin-terser';
+import {defineConfig}       from 'rollup';
 import {brotliCompressSync} from 'zlib';
 
 function wrapOutput() {
@@ -22,25 +24,32 @@ function wrapOutput() {
 }
 
 // eslint-disable-next-line arca/no-default-export
-export default [
-  {
-    input: `./sources/worker-zip/Worker.ts`,
-    output: {
-      file: `./sources/worker-zip/index.js`,
-      format: `cjs`,
-      strict: false,
-      preferConst: true,
-    },
-    plugins: [
-      resolve({
-        extensions: [`.mjs`, `.js`, `.ts`, `.tsx`, `.json`],
-        rootDir: path.join(__dirname, `../../`),
-        jail: path.join(__dirname, `../../`),
-        preferBuiltins: true,
-      }),
-      esbuild({tsconfig: false, target: `node12`}),
-      cjs({transformMixedEsModules: true, extensions: [`.js`, `.ts`]}),
-      wrapOutput(),
-    ],
+export default defineConfig({
+  input: `./sources/worker-zip/Worker.ts`,
+  output: {
+    file: `./sources/worker-zip/index.js`,
+    format: `cjs`,
+    strict: false,
+    generatedCode: `es2015`,
   },
-];
+  plugins: [
+    resolve({
+      extensions: [`.mjs`, `.js`, `.ts`, `.tsx`, `.json`],
+      rootDir: path.join(__dirname, `../../`),
+      jail: path.join(__dirname, `../../`),
+      preferBuiltins: true,
+    }),
+    esbuild({
+      tsconfig: false,
+      target: `node12`,
+      define: {
+        document: `undefined`,
+        XMLHttpRequest: `undefined`,
+        crypto: `undefined`,
+      },
+    }),
+    cjs({transformMixedEsModules: true, extensions: [`.js`, `.ts`]}),
+    terser({ecma: 2019}),
+    wrapOutput(),
+  ],
+});