+++ packages/yarnpkg-core/sources/LightReport.ts
@@ -1,11 +1,11 @@
-import {Writable}                from 'stream';
+import {Writable}                             from 'stream';
 
-import {Configuration}           from './Configuration';
-import {MessageName}             from './MessageName';
-import {Report, TimerOptions}    from './Report';
-import {formatNameWithHyperlink} from './StreamReport';
-import * as formatUtils          from './formatUtils';
-import {Locator}                 from './types';
+import {Configuration}                        from './Configuration';
+import {MessageName}                          from './MessageName';
+import {Report, SectionOptions, TimerOptions} from './Report';
+import {formatNameWithHyperlink}              from './StreamReport';
+import * as formatUtils                       from './formatUtils';
+import {Locator}                              from './types';
 
 export type LightReportOptions = {
   configuration: Configuration;
@@ -56,6 +56,14 @@ export class LightReport extends Report {
   }
 
   reportCacheMiss(locator: Locator) {
+  }
+
+  startSectionSync<T>(opts: SectionOptions, cb: () => T) {
+    return cb();
+  }
+
+  async startSectionPromise<T>(opts: SectionOptions, cb: () => Promise<T>) {
+    return await cb();
   }
 
   startTimerSync<T>(what: string, opts: TimerOptions, cb: () => T): T;