+++ CHANGELOG.md
@@ -10,12 +10,29 @@ Yarn now accepts sponsorships! Please give a look at our [OpenCollective](https:
 
 ### Installs
 
-- The pnpm linker will now remove the `node_modules/.store` and `node_modules` folders if they are empty.
 - The node-modules linker now tolerates if node_modules is a symbolic link and does not recreates it
+- The pnpm linker has received various improvements:
+  - It will now remove the `node_modules/.store` and `node_modules` folders if they are empty.
+  - It now supports running binaries of soft links.
+  - It will now create self-references for packages that don't depend on other versions of themselves.
+  - It will now remove scope folders (e.g. `node_modules/@yarnpkg`) if they are empty or after removing a scoped dependency.
 
+### Bugfixes
+
+- `@yarnpkg/pnpify` now escapes paths correctly
+- The ESM loader is now enabled regardless of the entrypoint module type, this fixes support for dynamic imports in commonjs modules when the entrypoint is also commonjs
+- `yarn workspaces foreach run` is now able to run binaries
+- The `node` field inside the `npm_config_user_agent` Yarn sets will now include a leading `v`.
+- Yarn is now able to recover from a corrupted install state.
+>
 ### Miscellaneous Features
 
 - Reporting for Git errors has been improved.
+- The resolution step now has a progress indicator.
+
+## 3.1.1
+
+- Updates the PnP compatibility layer for TypeScript 4.5
 
 ## 3.1.0